cleanup outdated core showcase
